+++ src/srt.ml
@@ -96,6 +96,7 @@ type (_, _) socket_opt =
   | Streamid : ([ `Read | `Write ], string) socket_opt
   | Passphrase : ([ `Write ], string) socket_opt
   | Pbkeylen : ([ `Read | `Write ], int) socket_opt
+  | Ipv6only : ([ `Read | `Write ], bool) socket_opt
 
 let messageapi = Messageapi
 let payloadsize = Payloadsize
@@ -116,6 +117,7 @@ let enforced_encryption = Enforced_encryption
 let passphrase = Passphrase
 let pbkeylen = Pbkeylen
 let streamid = Streamid
+let ipv6only = Ipv6only
 
 let srt_socket_opt_of_socket_opt (type a b) :
     (a, b) socket_opt -> Srt.socket_opt = function
@@ -138,6 +140,7 @@ let srt_socket_opt_of_socket_opt (type a b) :
   | Pbkeylen -> `Pbkeylen
   | Passphrase -> `Passphrase
   | Streamid -> `Streamid
+  | Ipv6only -> `Ipv6only
 
 let srtt_live = Int64.to_int srtt_live
 let srtt_file = Int64.to_int srtt_file
@@ -255,6 +258,7 @@ let getsockflag : type a b. socket -> (a, b) socket_opt -> b =
     | Transtype -> transtype_of_int !@arg
     | Passphrase -> to_string ()
     | Streamid -> to_string ()
+    | Ipv6only -> to_bool ()
 
 let setsockflag : type a b. socket -> (a, b) socket_opt -> b -> unit =
  fun sock opt v ->
@@ -293,6 +297,7 @@ let setsockflag : type a b. socket -> (a, b) socket_opt -> b -> unit =
           (f int transtype, sizeof int)
       | Passphrase -> of_string v
       | Streamid -> of_string v
+      | Ipv6only -> of_bool v
   in
   ignore
     (check_err (setsockflag sock (srt_socket_opt_of_socket_opt opt) arg arglen))