+++ src/srt.mli
@@ -58,6 +58,7 @@ val enforced_encryption : ([ `Write ], bool) socket_opt
 val streamid : ([ `Read | `Write ], string) socket_opt
 val passphrase : ([ `Write ], string) socket_opt
 val pbkeylen : ([ `Read | `Write ], int) socket_opt
+val ipv6only : ([ `Read | `Write ], bool) socket_opt
 
 type errno =
   [ `Easyncfail